Tree: Fix inherited type value modification
This fixes a problem where modifying the value of a symbol of a type
that is inherited from another type would fail during instantiation
with a `pymoca.tree.ModificationTargetNotFound`.

This also expands the "Processing failed for symbol" message given when
symbol type instantiation fails in `build_instance_tree`. It now
includes the message from the underlying exception to give a more
specific reason for the failure.

This is related to 4dd9048 that fixed inherited types for two test cases
using SI units in MSL 3 and 4, but they did not include testing value
*modification* of the inherited types.