Tree: Fix storing constant references in non-elementary typed equations
For the added test case (see the diff), what was happening is that we
added the resolved constant to the wrong class. We added it to the
Package.Length class, i.e. the InstanceClass of the type of
`distance_m`. Basically what was happening in the example was

1. Entering Model
2. Entering Length
3. Finding reference to constant, changing reference and storing new
   constant symbol
4. Exiting Length - popping off the stored constant symbols
5. Exiting Model

The reason for this happening was that it was not envisioned at the time
of the offending code that we would be entering InstanceClasses other
than those part of the lookup tree hierarchy.

Instead, we want to add the stored constant symbols to `Model`. So what
we currently implement for the last two steps above is:

4. Exiting Length - skipping popping
5. Exiting Model - popping off the stored constant symbols

Closes #301