WIP: Scope everywhere~
Scope is everywhere, but now how to flatten? ComponentRefs can be looked
up relative to the scope of a modification, but then... eh, that symbol
might change/be flattened/whatever. How does that work? Do we keep track
of what symbols go where? Do we do things in a certain order?