Update .git-blame-ignore-revs
The previous version of this file had commit hashes from the PR branch.
This version uses the merged commit hashes on master and adds another
commit from #289 that is only about formatting.

Lesson learned: This will have to be updated after any formatting-only
commits in a PR are merged into master/main.