CI/CD: Update max Python to 3.12
We need setuptools as a dependency for distutils.ccompiler to work in
Python 3.12 and higher.

Also since there are so many versions of Python active (3.8 to 3.12),
it seems excessive to test every version. This changes
the matrix to just three versions: 3.8, 3.10, 3.12. For future updates,
I envision picking min, max, and median of the supported Python
versions.