CI/CD: Use pre-commit for style checks
This changes to using the pre-commit GitHub action for style checks
that were previously run from tox. This allows configuration for
style checks locally and on GitHub to be consolidated into one place.

This uses the pre-commit GitHub action rather than the pre-commit.ci
service that requires install of a GitHub application for the repo.
The pre-commit.ci service claims to be faster and also allows
automatic fixing of black format as well as automatic updating of
tool versions in the config file, but I chose to not do this to avoid
having many bot-created commits in our repo vs just catching and fixing
these issues locally before pushing to remote. That seems cleaner at the
expense of a little more developer burden. It does depend on individual
developers installing the local tooling, which could be a problem with
larger numbers of contributors, so we may need to revisit.