Skip xml tests
The xml test failures are distracting in current development testing
when everything else passes. They were skipped in the
GitHub CI, so why not locally also?

The Modelica XML GitHub repo was last touched 8 years ago so I propose
deprecating it in a future release unless someone wants to fix it.