WIP: MODELICAPATH tree as stub `ast.Class`es
First step WIP for integrating MODELICAPATH with new name lookup in
a DRY way. This implements MODELICAPATH capability that works with the
current name lookup and should integrate easily into the new. It also
works with the new parse caching.

It adds `parsed` and `path` attributes to `ast.Class` for lazy parsing.
Calling these "stub classes" since it is just a Class to be
filled in later when it is parsed. Creates tree of stub classes for all
directories and class.mo files in MODELICAPATH.

Added a `LazyParseDict` class to do the lazy parse when class
is accessed through the `ast.Class.classes` dictionary in name lookup.

Added `parse_text()` that returns a `parser.ModelicaFile` so that the
lazily-parsed class can be promptly inserted into the tree that already
exists. The previous `parse()` is still there but returns a tree where
the parsed class is several levels down if there is a `within` statement
in the Modelica text.

The added test case is very basic, but shows it works with existing
name lookup.

TODO: Add more test cases (and fix) until coverage is decent.
TODO: Most likely some refactoring since this is a prototype.