parse_test/cache: Check log messages cache was hit
When running on Github Actions, the first and second hit time have a
tendency to be equal, possibly for reasons related to mounted/network
storage.

To make sure that the cache is hit, we can also check whether the log
messages indicate as such, instead of just looking at a forced update of
the last hit time.

Note that this commit does not fix the failing of the test, it just adds
an additional check to better understand what is going on.