CI/CD: Fix failing cache test likely due to SQLite sync issues
On GitHub Actions, and only on the Windows platform, the check
`self.assertGreater(second_hit_time, first_hit_time)` would fail, as
both times would be equal. There is probably some funky stuff going on
with SQLite not being able to properly notify the reader (the test) that
a writer (the `parse` call) made changes. We therefore force a close and
re-open of the database.

Note that:
- The second parse call does correctly read from the cache. We check the
  log messages that this is the case (see parent commit)
- Such erratic reader/writer sync behavior is probably because of the
  use of mounted storage.