setup: Fully switch from setup.py to pyproject.toml
We were partially using both, which is asking for trouble when it comes
to various build tools.

We split off the antlr compile command to its separate file (not part of
the pymoca library). It's just a `python antlr_build.py` call now,
instead of having `python setup.py antlr`.