CI/CD: Fix failing cache test due to coarse clock on Windows
We compare the `last_prune` / `last_hit` times in the tests, checking
that a second successive call updates the value. The time between calls
might take a couple hundred us to a few ms, especially when there is
very little work we do inbetween like in `test_parse_cache_hit`.

This particular `test_parse_cache_hit` test was also the test that fails
on Windows on CI/CD, as it turns out that the clock on that platform is
quite coarse. Two successive calls to `parser.parse()` meaning two
successive calls to `_microseconds_since_epoch()`, would return the
exact same time twice. This would cause the assertion in the test to
fail.

There are more accurate clocks available ( see `time.time_ns()`,
https://peps.python.org/pep-0564/ ), but even those are not enough on
Windows to make the tests succeed at all times.

As a workaround we therefore make sure that, when updating the
`last_hit` / `last_prune` time, we at the very least increment it by 1
(microsecond).