parser/cache: Refactor _microseconds_since_epoch by passing timedelta
Instead of relying on the "datetime.utcnow()" and the subtracting of
timedeltas, we know that we are only interested in relative offset
compared to the current. We can therefore simplify the code by making
`_microseconds_since_epoch` accept a timedelta instead.

This paves the way for usage of `time.time_ns()` as a more accurate
clock, and solving the deprecation of `datetime.utcnow()` in the
process. This will be addressed in a next commit.