examples/Spring: Fix wrong keyword to control.locus kist->kvect
We should call control.rlocus with `kvect` rather than `klist` because
the latter does not work in the currently most recent version of control
(=0.10.0). The future rename of `kvect` to `gains` isn't yet available
in this version, so we explicitly pin it for now to avoid another
possible regression in running this example down the line when `kvect`
is deprecated.

Closes #324