Tree: Move obsolete find_name args to stub code
Remove the arguments to find_name that are now obsolete with the new
instantiation and move them to the stub code in ast.Class that patches
in find_name to replace ast.Class.find_class. That will also eventually
be removed if find_name becomes the production name lookup code.