Tree: Add simple name lookup
Add name lookup for the the "A" part of a name "A.B.C". It tries to
closely follow the Modelica Language Specification v3.5, but there are
several TODO for yet to be implemented features of full name lookup.

Add a specific `NameLookupError` exception.

We also add tests cases from ModelicaCompliance for "Simple Name Lookup"
as defined in the Modelica Language Specification.