Tree: Fix and optimize lookup of extends clauses in other extends nodes
The specification is a little vague on this, but the compliance test
`ModelicaCompliance\Inheritance\Flattening\InheritedBaseClass.mo` states
that "[...] a name used in an extends-clause may not be inherited".

We would not catch this is an all possible cases (see added tests).
Furthermore we would perform a lookup of e.g. "extends A" inside the
found "A" itself (and other extends nodes). This is rather costly,
especially as finding a class there would not be allowed anyway. To this
end, we introduce the `search_inherited=False` option to `_find_name`.

Lastly, we add a specific check and error message for this rule, such
that users know what is wrong with their models instead of a message
about a "class not found", "cannot extend symbol", or some more obscure
error down the line.