Tree: Optimize partial instantiation of parents
The previous implementation before this commit would perform an `extend`
call on every level up, which in turn would call `update_parent_refs`.
Both `extend` and `update_parent_refs` would traverse the tree all the
down from root.

We can optimize this by only doing these calls _once_, when we are done
and have traversed all the way up to root. Note that that still means we
traverse the trees downwards twice.

Note that it also means we do these operations of tree merging every
time we find an ast.Class instead of an tree.InstanceClass in the name
lookup. An alternative to avoid this is to partially instantiate a class
when we go down a level in the `find_name` call. This would require
access to `InstanceClass`, which these functions do not currently have.