Parser/AST: Add visibility to `Class`
This adds the visibility attribute to `Class` like was already in
`Symbol` and `ExtendsClause`, so we now track `public` and `protected`
visibility keywords for all necessary elements in the composition rule
in the Modelica grammar. Visibility determines if a symbol declaration
or class definition is visible outside of the class where it is defined.
Protected class definitions are less common than protected symbols, but
there are some in the Modelica Standard Library.

A class can have multiple, interleaved public and private sections. This
commit fixes an issue where only the last of the `public` and
`protected` sections were kept in a composition. We now keep a list of
element lists for all of the private sections and do the same for all of
the public sections.

Yet to be implemented is full error checking on visibility. See, for
instance, section 4.1 of the spec. We do have a check on visibility for
imported names in name lookup.