CI/CD Fix deploy with no `setup.py`
Fix pymoca package build in the GitHub CI "deploy" job so that it works
with `setup.py` removed (which was done back in commit 329bc18). I guess
we have not done a deploy since we fully moved to `pyproject.toml`, so
this is the first time we are seeing the issue. Before these changes,
the package version was 0.0.0. After these changes, `python -m build` is
picking up the git tag correctly:
```
Successfully built pymoca-0.11.0.dev2.tar.gz and pymoca-0.11.0.dev2-py3-none-any.whl
```
(We now use `python -m build` instead of `python setup.py sdist
bdist_wheel` to build the package.)

The combination of changes in this commit work, but I'm not super
confident that it will still work in future versions of setuptools
and/or versioneer. The PyPI documentation for versioneer 0.29 (the
latest version) says it is compatible with with Python up to 3.11, but
it still appears to work OK with 3.13 that we are now supporting with
Pymoca.

I also found that including a minimal `setup.py` such as the
following works instead of the change to `pyproject.toml` in this commit:

```python
from setuptools import setup

import versioneer

setup(
    version=versioneer.get_version(),
    cmdclass=versioneer.get_cmdclass(),
)
```

Adding `setup.py` back in could be a fall-back solution if we find that
this commit fails in future versions of setuptools and/or versioneer.
The current versioneer README documentation and some of the issues on
the versioneer GitHub repo suggest that `setup.py` is still needed, but
the solution in this commit is from the currently open GitHub issue
referenced below.

References:
- https://github.com/python-versioneer/python-versioneer/issues/370
- https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
- https://setuptools.pypa.io/en/latest/build_meta.html