Remove ANTLR files from `MANIFEST.in`
Remove the `java` directory from `MANIFEST.in`. These are the files
that are included into the distribution. This directory does not exist
and produced a warning to that effect during the distribution build.

The `java` directory was previously renamed to `antlr` in the
repository and currently includes a script and an ANTLR jar file used
to generate the parser. We expect that developers wishing to change the
grammar and parser will want to do so from the repository and not the
distribution, so we remove it the ANTLR directory instead of fixing it in `MANIFEST.in`.