CI/CD: Include xml schema files in distribution
Include submodules in the deploy step so that the build process can
find the xml schema files specified in `MANIFEST.in`. Previous
distributions were apparently missing these files which I think means
the xml backend would not work in the distribution. I'm not going
to check because we don't actively develop the xml backend. However
I thought we would go ahead and fix this issue to remove the warning
from the build process.