Parser: Update ANTLR to 4.13.1
Our ANTLR 4.7.* runtime uses `typing.io` that has been deprecated in
Python 3.12 release, so we are updating to the ANTLR version that we run
on `master` at this time.

See 65ff7987 (the commit on master) for more details. Note that that
commit does not cleanly apply, as we are missing other changes to the
grammar on this maintenance branch.