WIP: Add extends indirect modification test
Adds `test_extends_encapsulated_class_indirect_modification` to test
cases where there are extends class_modifications that are within a
name lookup path.

This is a separate commit from the ones that fix the issue because
we are trying a couple of different approaches to fix the issue.

TODO: Squash into the selected fix commit.