Tree: Fix missed InstanceClass attributes init
Add initialization of the `encapsulated` and `partial` attributes of
`InstanceClass`. This fixes a problem where the `encapsulated=True`
value was lost during instantiation which caused name lookup to fail. We
preemptively fix the missing `partial` attribute while we are here and
also clean up two other initializers to be like others done during
object creation instead of after.