Tree: Instantiate using find_name_path
Fix instantiation to utilize the new find_name_path method and apply
modifications to children in the path.

Instances are stored in the passed-in parent instance node or contained
symbol instance, which is often not the parent node from the AST class
definition. The start of the process follows the spec; the top-level
instantiated class parent is the root of the instance tree. (We
do no always add instances to the classes lists according to the AST
structure because we need to allow for multiple instances of the same
name with possibly different modifications.)

Extends clauses are included in the name lookup path and are now
instantiated after lookup using that information, including modifiers.
To enable this, parent is added to ExtendsClause creation in the parser
and we add an _instantiate_path_partially method to pass down
modifications to the children in the path before fully instantiating the
unnamed extends node.

We similarly use _instantiate_path_partially for symbol types.

We separated out _apply_modifications from _instantiate_partially to
allow it to be used elsewhere. We are currently only using it from
_instantiate_partially, but we keep it separate to reduce function size
and complexity.

The test_instantiation_tree test in parse_test.py is updated:
- Adds modifiers in the InstantiationTree.mo test model to check for
  proper modification handling.
- Changes the overall test focus from one that checked that all
  instances were added to Class.classes lists to one that checks that
  the instance is properly instantiated with correct modifications
  according to the spec in preparation for flattening.